+++ app/package.json
@@ -10,16 +10,18 @@
   },
   "dependencies": {
     "@react-three/fiber": "^8.15.12",
-    "connectkit": "^1.5.3",
+    "@tanstack/query-core": "^5.22.2",
+    "@tanstack/react-query": "^5.22.2",
+    "circuits": "1.0.0",
+    "connectkit": "^1.7.1",
     "ethers": "5.7.2",
     "next": "14.0.3",
     "react": "^18",
     "react-beautiful-dnd": "^13.1.1",
     "react-dom": "^18",
     "three": "^0.159.0",
-    "viem": "^1.21.4",
-    "wagmi": "^1.4.13",
-    "circuits": "1.0.0"
+    "viem": "^2.7.13",
+    "wagmi": "^2.5.7"
   },
   "devDependencies": {
     "@types/node": "^20",
